Feat: expose jointType as an enum dropdown (PMX JointType fields)
Convert the pmxPhysicsNode joints[] jointType child from a plain long to
an MFnEnumAttribute, matching the bodyGroupId/bodyColliderType/
bodyPhysicsMode pattern: one field per PMX JointType (Spring6Dof/SixDof/
P2P/ConeTwist/Slider/Hinge, values 0..5), so the Attribute Editor shows a
dropdown instead of a raw integer.  The command writes it via setShort
(like bodyGroupId); readJointData / config signature read it via asInt,
unchanged.  Storable/keyable set in the enum block (kept out of the
numeric loop).

- jointType is an enum dropdown in the Attribute Editor.
- New cmd test: attributeQuery(listEnum) exposes the six fields; writing
  type=4 (Slider) round-trips as 4.
- cmd suite 51 passed (36 body + 15 constraint), import + node pass.

+++ mmd/maya/cmds/rigid_body_constraint_cmd.cpp
@@ -320,7 +320,9 @@ MStatus doCreate(const MArgParser& parser, const MObject& solverNode, int& outIn
     elem.child(PhysicsNode::aJointNameUniversal).setString(nameUniversal);
     elem.child(PhysicsNode::aJointBodyA).setInt(bodyA);
     elem.child(PhysicsNode::aJointBodyB).setInt(bodyB);
-    elem.child(PhysicsNode::aJointType).setInt(type);
+    // jointType is an enum attribute — write through setShort (like the
+    // bodyGroupId enum).
+    elem.child(PhysicsNode::aJointType).setShort(static_cast<short>(type));
     mmd::maya::setPlugDouble3(elem.child(PhysicsNode::aJointFrameTranslate), localT);
     mmd::maya::setPlugDouble3(elem.child(PhysicsNode::aJointFrameRotate), localR);
     mmd::maya::setPlugDouble3(elem.child(PhysicsNode::aJointLinearMin), lmin);
